producer/requirements.txt
flask==1.1.4
werkzeug==1.0.1
kafka-python==2.0.2
pymongo[srv]==3.12.2
pymongo==3.12.2
jinja2==2.11.3
markupsafe==1.1.1
